Moved Shakespeare model
