+++ accdfl/core/model/__init__.py
@@ -9,7 +9,7 @@ def serialize_model(model: torch.nn.Module) -> bytes:
 
 def create_model(dataset: str):
     if dataset in ["shakespeare", "shakespeare_sub", "shakespeare_sub96"]:
-        from accdfl.core.datasets.Shakespeare import LSTM
+        from accdfl.core.models.shakespeare import LSTM
         return LSTM()
     elif dataset == "cifar10" or dataset == "cifar10_niid":
         from accdfl.core.models.cifar10 import GNLeNet
